adding mongodb

+++ src/routes/posts-router.ts
@@ -4,17 +4,17 @@ import { basicAuthMiddleware } from "../middleware/auth/basic";
 import { createUpdateBodyValidationMiddleware } from "../middleware/validation/validation-posts";
 import { sendErrorsIfAnyMiddleware } from "../middleware/validation/validation-universal";
 import { Post } from "../repositories/posts-repo";
-import { v4 as uuidv4 } from "uuid";
 
 const router = express.Router();
 
-router.get("/", (req, res) => {
-  res.send(postsRepository.findAll());
+router.get("/", async (req, res) => {
+  const posts = await postsRepository.findAll();
+  res.send(posts);
 });
 
-router.get("/:id", (req, res) => {
+router.get("/:id", async (req, res) => {
   const id = req.params.id;
-  const post = postsRepository.findById(id);
+  const post = await postsRepository.findById(id);
   if (post) {
     res.status(200).json(post);
   } else {
@@ -28,21 +28,21 @@ router.post(
   "/",
   createUpdateBodyValidationMiddleware,
   sendErrorsIfAnyMiddleware,
-  (req: Request, res: Response) => {
+  async (req: Request, res: Response) => {
     const { title, shortDescription, content, blogId, blogName } = req.body;
 
-    const id = uuidv4();
-
     const post: Post = {
-      id,
       title,
       shortDescription,
       content,
       blogId,
       blogName,
     };
-    postsRepository.create(post);
-    return res.status(201).json(postsRepository.findById(id));
+    const insertedId = await postsRepository.create(post);
+
+    const createdPost = await postsRepository.findById(insertedId);
+
+    return res.status(201).json(createdPost);
   }
 );
 
@@ -50,12 +50,11 @@ router.put(
   "/:id",
   createUpdateBodyValidationMiddleware,
   sendErrorsIfAnyMiddleware,
-  (req: Request, res: Response) => {
+  async (req: Request, res: Response) => {
     const id: string = req.params.id;
     const { title, shortDescription, content, blogId, blogName } = req.body;
 
-    const updated = postsRepository.updateById(id, {
-      id,
+    const isUpdated = await postsRepository.updateById(id, {
       title,
       shortDescription,
       content,
@@ -63,7 +62,7 @@ router.put(
       blogName,
     });
 
-    if (updated) {
+    if (isUpdated) {
       res.sendStatus(204);
     } else {
       res.sendStatus(404);
@@ -71,13 +70,19 @@ router.put(
   }
 );
 
-router.delete("/:id", (req, res) => {
+router.delete("/:id", async (req, res) => {
   const id = req.params.id;
-  const deleted = postsRepository.deleteById(id);
-  if (deleted) {
-    res.sendStatus(204);
+  const post = await postsRepository.findById(id);
+  if (!post) {
+    return res.sendStatus(404);
+  }
+
+  const isDeleted = await postsRepository.deleteById(id);
+
+  if (isDeleted) {
+    return res.sendStatus(204);
   } else {
-    res.sendStatus(404);
+    return res.sendStatus(500);
   }
 });
